fix: moved the code to the component
